# internal/interfaces/http/handlers/handlers.go
package handlers

import (
	"sync"
	"time"
	"vibe-storm/internal/application/dto"
	"vibe-storm/pkg/config"
	"vibe-storm/pkg/middleware"

	"github.com/go-playground/validator/v10"
	"github.com/labstack/echo/v4"
	"github.com/sirupsen/logrus"
	"gorm.io/gorm"
)

// Handler dependencies that can be injected
type HandlerDeps struct {
	DB     *gorm.DB
	Config *config.Config
}

// Base handler interface
type Handler[T any] interface {
	Handle(c echo.Context) (T, error)
}

// HealthCheckHandler handles health check requests
type HealthCheckHandler struct {
	Deps HandlerDeps
}

// Handle implements the health check logic
//
//	@Summary		Health Check
//	@Description	Get the health status of the application
//	@Tags			health
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	dto.HealthResponse	"Health status"
//	@Router			/health [get]
func (h *HealthCheckHandler) Handle(c echo.Context) (dto.HealthResponse, error) {
	return dto.HealthResponse{
		Status:  "healthy",
		Service: h.Deps.Config.App.Name,
		Version: h.Deps.Config.App.Version,
		Env:     h.Deps.Config.App.Env,
	}, nil
}

// HomePageHandler handles home page requests
type HomePageHandler struct {
	Deps HandlerDeps
}

// Handle implements the home page logic
//
//	@Summary		Home Page
//	@Description	Get the home page information
//	@Tags			general
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	dto.HomeResponse	"Home page info"
//	@Router			/ [get]
func (h *HomePageHandler) Handle(c echo.Context) (dto.HomeResponse, error) {
	return dto.HomeResponse{
		Message: "Welcome to " + h.Deps.Config.App.Name,
		Version: h.Deps.Config.App.Version,
		Status:  "running",
	}, nil
}

// CreateUserHandler handles user creation requests
type CreateUserHandler struct {
	Deps HandlerDeps
}

// Handle implements the user creation logic
//
//	@Summary		Create User
//	@Description	Create a new user account
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			user	body		dto.CreateUserRequest	true	"User data"
//	@Success		201		{object}	dto.UserResponse		"Created user"
//	@Failure		400		{object}	dto.ErrorResponse		"Invalid request data"
//	@Failure		409		{object}	dto.ErrorResponse		"User already exists"
//	@Failure		500		{object}	dto.ErrorResponse		"Internal server error"
//	@Router			/users [post]
//	@Security		ApiKeyAuth
func (h *CreateUserHandler) Handle(c echo.Context) (dto.UserResponse, error) {
	var req dto.CreateUserRequest
	if err := c.Bind(&req); err != nil {
		return dto.UserResponse{}, &dto.StructuredError{
			Response: dto.NewStructuredErrorResponse(
				dto.ErrCodeInvalidRequest,
				c.Request().URL.Path,
			),
		}
	}

	// Validate the request DTO
	if err := dto.ValidateStruct(&req); err != nil {
		validationErrors, ok := err.(validator.ValidationErrors)
		if !ok {
			return dto.UserResponse{}, &dto.StructuredError{
				Response: dto.NewStructuredErrorResponse(
					dto.ErrCodeValidationError,
					c.Request().URL.Path,
				),
			}
		}
		return dto.UserResponse{}, &dto.StructuredError{
			Response: dto.NewValidationErrorResponse(
				c.Request().URL.Path,
				validationErrors,
			),
		}
	}

	// TODO: Implement user creation logic
	return dto.UserResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}

// GetUsersHandler handles get users requests
type GetUsersHandler struct {
	Deps HandlerDeps
}

// Handle implements the get users logic
//
//	@Summary		Get Users
//	@Description	Get a list of all users with pagination and filtering
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			query	query		dto.GetUsersQuery	false	"Query parameters"
//	@Success		200		{object}	dto.GetUsersResponse	"List of users with pagination"
//	@Failure		400		{object}	dto.ErrorResponse		"Invalid query parameters"
//	@Failure		500		{object}	dto.ErrorResponse		"Internal server error"
//	@Router			/users [get]
//	@Security		ApiKeyAuth
func (h *GetUsersHandler) Handle(c echo.Context) (dto.GetUsersResponse, error) {
	var query dto.GetUsersQuery
	if err := c.Bind(&query); err != nil {
		return dto.GetUsersResponse{}, &dto.StructuredError{
			Response: dto.NewStructuredErrorResponse(
				dto.ErrCodeInvalidParameter,
				c.Request().URL.Path,
			),
		}
	}

	// Validate query parameters
	if err := dto.ValidateStruct(&query); err != nil {
		validationErrors, ok := err.(validator.ValidationErrors)
		if !ok {
			return dto.GetUsersResponse{}, &dto.StructuredError{
				Response: dto.NewStructuredErrorResponse(
					dto.ErrCodeValidationError,
					c.Request().URL.Path,
				),
			}
		}
		return dto.GetUsersResponse{}, &dto.StructuredError{
			Response: dto.NewValidationErrorResponse(
				c.Request().URL.Path,
				validationErrors,
			),
		}
	}

	// TODO: Implement get users logic with query parameters
	return dto.GetUsersResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}

// GetUserHandler handles get user by ID requests
type GetUserHandler struct {
	Deps HandlerDeps
}

// Handle implements the get user by ID logic
//
//	@Summary		Get User
//	@Description	Get a user by their ID
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id	path		string			true	"User ID"
//	@Success		200	{object}	dto.UserResponse	"User data"
//	@Failure		404	{object}	dto.ErrorResponse	"User not found"
//	@Failure		500	{object}	dto.ErrorResponse	"Internal server error"
//	@Router			/users/{id} [get]
//	@Security		ApiKeyAuth
func (h *GetUserHandler) Handle(c echo.Context) (dto.UserResponse, error) {
	// TODO: Implement get user by ID logic
	return dto.UserResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}

// UpdateUserHandler handles user update requests
type UpdateUserHandler struct {
	Deps HandlerDeps
}

// Handle implements the user update logic
//
//	@Summary		Update User
//	@Description	Update a user by their ID
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id	path		string				true	"User ID"
//	@Param			user	body		dto.UpdateUserRequest	true	"Updated user data"
//	@Success		200		{object}	dto.UserResponse		"Updated user"
//	@Failure		400		{object}	dto.ErrorResponse		"Invalid request data"
//	@Failure		404		{object}	dto.ErrorResponse		"User not found"
//	@Failure		500		{object}	dto.ErrorResponse		"Internal server error"
//	@Router			/users/{id} [put]
//	@Security		ApiKeyAuth
func (h *UpdateUserHandler) Handle(c echo.Context) (dto.UserResponse, error) {
	// TODO: Implement update user logic
	return dto.UserResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}

// DeleteUserHandler handles user deletion requests
type DeleteUserHandler struct {
	Deps HandlerDeps
}

// Handle implements the user deletion logic
//
//	@Summary		Delete User
//	@Description	Delete a user by their ID
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id	path	string		true	"User ID"
//	@Success		204	"No Content"
//	@Failure		404	{object}	dto.ErrorResponse	"User not found"
//	@Failure		500		{object}	dto.ErrorResponse		"Internal server error"
//	@Router			/users/{id} [delete]
//	@Security		ApiKeyAuth
func (h *DeleteUserHandler) Handle(c echo.Context) (dto.DeleteUserResponse, error) {
	// TODO: Implement delete user logic
	return dto.DeleteUserResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}

// RateLimiter for brute force protection
type RateLimiter struct {
	attempts    map[string][]time.Time
	mutex       sync.RWMutex
	maxAttempts int
	window      time.Duration
}

// NewRateLimiter creates a new rate limiter
func NewRateLimiter(maxAttempts int, window time.Duration) *RateLimiter {
	return &RateLimiter{
		attempts:    make(map[string][]time.Time),
		maxAttempts: maxAttempts,
		window:      window,
	}
}

// IsBlocked checks if an IP is blocked due to too many failed attempts
func (rl *RateLimiter) IsBlocked(ip string) bool {
	rl.mutex.Lock()
	defer rl.mutex.Unlock()

	now := time.Now()
	var validAttempts []time.Time

	// Filter out old attempts
	for _, attempt := range rl.attempts[ip] {
		if now.Sub(attempt) < rl.window {
			validAttempts = append(validAttempts, attempt)
		}
	}

	rl.attempts[ip] = validAttempts
	return len(validAttempts) >= rl.maxAttempts
}

// RecordFailedAttempt records a failed login attempt
func (rl *RateLimiter) RecordFailedAttempt(ip string) {
	rl.mutex.Lock()
	defer rl.mutex.Unlock()

	now := time.Now()
	rl.attempts[ip] = append(rl.attempts[ip], now)
}

// Reset clears failed attempts for an IP (on successful login)
func (rl *RateLimiter) Reset(ip string) {
	rl.mutex.Lock()
	defer rl.mutex.Unlock()

	delete(rl.attempts, ip)
}

// Global rate limiter instance for signin protection
var signinRateLimiter = NewRateLimiter(5, 15*time.Minute) // 5 attempts per 15 minutes

// SignupHandler handles user registration requests
type SignupHandler struct {
	Deps HandlerDeps
}

// Handle implements the user signup logic
//
//	@Summary		User Signup
//	@Description	Register a new user account
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			user	body		dto.SignupRequest	true	"User signup data"
//	@Success		201		{object}	dto.AuthResponse		"Authentication tokens and user data"
//	@Failure		400		{object}	dto.ErrorResponse		"Invalid request data"
//	@Failure		409		{object}	dto.ErrorResponse		"User already exists"
//	@Failure		500		{object}	dto.ErrorResponse		"Internal server error"
//	@Router			/auth/signup [post]
func (h *SignupHandler) Handle(c echo.Context) (dto.AuthResponse, error) {
	var req dto.SignupRequest
	if err := c.Bind(&req); err != nil {
		return dto.AuthResponse{}, &dto.StructuredError{
			Response: dto.NewStructuredErrorResponse(
				dto.ErrCodeInvalidRequest,
				c.Request().URL.Path,
			),
		}
	}

	// Validate the request DTO
	if err := dto.ValidateStruct(&req); err != nil {
		validationErrors, ok := err.(validator.ValidationErrors)
		if !ok {
			return dto.AuthResponse{}, &dto.StructuredError{
				Response: dto.NewStructuredErrorResponse(
					dto.ErrCodeValidationError,
					c.Request().URL.Path,
				),
			}
		}
		return dto.AuthResponse{}, &dto.StructuredError{
			Response: dto.NewValidationErrorResponse(
				c.Request().URL.Path,
				validationErrors,
			),
		}
	}

	// TODO: Implement user creation and token generation logic
	// For now, return placeholder response
	return dto.AuthResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}

// SigninHandler handles user authentication requests with brute force protection
type SigninHandler struct {
	Deps HandlerDeps
}

// Handle implements the user signin logic
//
//	@Summary		User Signin
//	@Description	Authenticate user and return access tokens
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			user	body		dto.SigninRequest	true	"User signin credentials"
//	@Success		200		{object}	dto.AuthResponse		"Authentication tokens and user data"
//	@Failure		400		{object}	dto.ErrorResponse		"Invalid request data"
//	@Failure		401		{object}	dto.ErrorResponse		"Invalid credentials"
//	@Failure		429		{object}	dto.ErrorResponse		"Too many failed attempts"
//	@Failure		500		{object}	dto.ErrorResponse		"Internal server error"
//	@Router			/auth/signin [post]
func (h *SigninHandler) Handle(c echo.Context) (dto.AuthResponse, error) {
	// Get client IP for rate limiting
	clientIP := c.RealIP()

	// Check if IP is blocked due to too many failed attempts
	if signinRateLimiter.IsBlocked(clientIP) {
		logrus.WithField("ip", clientIP).Warn("Blocked signin attempt due to rate limiting")
		return dto.AuthResponse{}, &dto.StructuredError{
			Response: dto.NewStructuredErrorResponse(
				dto.ErrCodeTooManyRequests,
				c.Request().URL.Path,
			),
		}
	}

	var req dto.SigninRequest
	if err := c.Bind(&req); err != nil {
		return dto.AuthResponse{}, &dto.StructuredError{
			Response: dto.NewStructuredErrorResponse(
				dto.ErrCodeInvalidRequest,
				c.Request().URL.Path,
			),
		}
	}

	// Validate the request DTO
	if err := dto.ValidateStruct(&req); err != nil {
		validationErrors, ok := err.(validator.ValidationErrors)
		if !ok {
			return dto.AuthResponse{}, &dto.StructuredError{
				Response: dto.NewStructuredErrorResponse(
					dto.ErrCodeValidationError,
					c.Request().URL.Path,
				),
			}
		}
		return dto.AuthResponse{}, &dto.StructuredError{
			Response: dto.NewValidationErrorResponse(
				c.Request().URL.Path,
				validationErrors,
			),
		}
	}

	// TODO: Implement user authentication logic
	// For now, simulate authentication failure for demonstration
	signinRateLimiter.RecordFailedAttempt(clientIP)

	return dto.AuthResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeUnauthorized,
			c.Request().URL.Path,
		),
	}
}

// RefreshTokenHandler handles token refresh requests
type RefreshTokenHandler struct {
	Deps HandlerDeps
}

// Handle implements the token refresh logic
//
//	@Summary		Refresh Token
//	@Description	Refresh access token using refresh token
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			token	body		dto.RefreshTokenRequest	true	"Refresh token data"
//	@Success		200		{object}	dto.TokenResponse		"New access token"
//	@Failure		400		{object}	dto.ErrorResponse		"Invalid request data"
//	@Failure		401		{object}	dto.ErrorResponse		"Invalid refresh token"
//	@Failure		500		{object}	dto.ErrorResponse		"Internal server error"
//	@Router			/auth/refresh [post]
func (h *RefreshTokenHandler) Handle(c echo.Context) (dto.TokenResponse, error) {
	var req dto.RefreshTokenRequest
	if err := c.Bind(&req); err != nil {
		return dto.TokenResponse{}, &dto.StructuredError{
			Response: dto.NewStructuredErrorResponse(
				dto.ErrCodeInvalidRequest,
				c.Request().URL.Path,
			),
		}
	}

	// Validate the request DTO
	if err := dto.ValidateStruct(&req); err != nil {
		validationErrors, ok := err.(validator.ValidationErrors)
		if !ok {
			return dto.TokenResponse{}, &dto.StructuredError{
				Response: dto.NewStructuredErrorResponse(
					dto.ErrCodeValidationError,
					c.Request().URL.Path,
				),
			}
		}
		return dto.TokenResponse{}, &dto.StructuredError{
			Response: dto.NewValidationErrorResponse(
				c.Request().URL.Path,
				validationErrors,
			),
		}
	}

	// TODO: Implement token refresh logic
	return dto.TokenResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}

// MeHandler handles get current user requests
type MeHandler struct {
	Deps HandlerDeps
}

// Handle implements the get current user logic
//
//	@Summary		Get Current User
//	@Description	Get the currently authenticated user's information
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	dto.UserResponse	"Current user data"
//	@Failure		401	{object}	dto.ErrorResponse	"Unauthorized"
//	@Failure		500	{object}	dto.ErrorResponse	"Internal server error"
//	@Router			/auth/me [get]
//	@Security		ApiKeyAuth
func (h *MeHandler) Handle(c echo.Context) (dto.UserResponse, error) {
	// Get user from JWT context
	_, ok := middleware.GetUserFromContext(c)
	if !ok {
		return dto.UserResponse{}, &dto.StructuredError{
			Response: dto.NewStructuredErrorResponse(
				dto.ErrCodeUnauthorized,
				c.Request().URL.Path,
			),
		}
	}

	// TODO: Implement get current user logic using userClaims.UserID
	return dto.UserResponse{}, &dto.StructuredError{
		Response: dto.NewStructuredErrorResponse(
			dto.ErrCodeInternalError,
			c.Request().URL.Path,
		),
	}
}